V2.3 beta
特征匹配的简单实现，替换原本的FeatureAssociation，将坐标结构体移动到baseFunc中，新增config功能，main_on_windows中路径使用增量式表达，改善内存回收问题（Disigned by Pyrokine）

+++ LSD/myLSD.h
@@ -129,7 +129,7 @@ namespace mylsd {
 	int Comp(const void *p1, const void *p2);
 
 	Mat createMapCache(Mat MapGray, double res, double z_occ_max_dis);
-	structLSD myLineSegmentDetector(Mat MapGray, int oriMapCol, int oriMapRow, double sca, double sig, double angThre, double denThre, double pseBin);
+	structLSD myLineSegmentDetector(Mat MapGray, int oriMapCol, int oriMapRow, double sca, double sig, double angThre, double denThre, int pseBin);
 	Mat GaussianSampler(Mat image, double sca, double sig);
 	structRegionGrower RegionGrower(int x, int y, Mat banMap, double regDeg, Mat degMap, double degThre);
 	structRec RectangleConverter(structReg reg, Mat magMap, double aliPro, double degThre);
